Modifying to run tests as part of the initial build process to determine whether or not to proceed

+++ scripts/docker/create_and_run_dockers.sh
@@ -34,6 +34,11 @@ for line in $(cat ${projfile}); do
         docker build -t ${image} - < ${modifiedslug}_Dockerfile
     fi
 
-    # Run the Docker image
-    docker run -it --rm -v $(pwd):/Scratch ${image} /bin/bash /home/awshi2/dt-fixing-tools/scripts/docker/run_experiment.sh ${slug}
+    # Run the Docker image if it exists
+    docker inspect ${image} > /dev/null 2>&1
+    if [ $? == 1 ]; then
+        echo "${image} NOT BUILT PROPERLY, LIKELY TESTS FAILED"
+    else
+        docker run -it --rm -v $(pwd):/Scratch ${image} /bin/bash /home/awshi2/dt-fixing-tools/scripts/docker/run_experiment.sh ${slug}
+    fi
 done
